Add test to toggle
