+++ src/addRemove.test.js
@@ -1,36 +1,9 @@
 import { saveList, addTaskListener, clearCompletedListener } from './crud';
 import ToDoList, { LIST_STORAGE_KEY } from './toDoList';
+import {eventMock} from './mock.js'
 
-const eventMock = {
-  preventDefault: () => {},
-  target: {
-    getElementsByTagName: () => [{ value: 'Mock description' }],
-  },
-};
 
-class LocalStorageMock {
-  constructor() {
-    this.store = {};
-  }
 
-  clear() {
-    this.store = {};
-  }
-
-  getItem(key) {
-    return this.store[key] || null;
-  }
-
-  setItem(key, value) {
-    this.store[key] = value;
-  }
-
-  removeItem(key) {
-    delete this.store[key];
-  }
-}
-
-global.localStorage = new LocalStorageMock();
 
 describe('Add Task', () => {
   let emptyList = new ToDoList();
