+++ src/mock.js
@@ -1,0 +1,30 @@
+export const eventMock = {
+  preventDefault: () => {},
+  target: {
+    getElementsByTagName: () => [{ value: 'Mock description' }],
+  },
+};
+
+class LocalStorageMock {
+  constructor() {
+    this.store = {};
+  }
+
+  clear() {
+    this.store = {};
+  }
+
+  getItem(key) {
+    return this.store[key] || null;
+  }
+
+  setItem(key, value) {
+    this.store[key] = value;
+  }
+
+  removeItem(key) {
+    delete this.store[key];
+  }
+}
+
+global.localStorage = new LocalStorageMock();
